SE-Attention added
This is an update to this project, this update content is: on the basis of the original algorithm, the attention mechanism is added, the attention mechanism uses the Squeeze-and-Excitation method, I have done the corresponding test locally, due to the selection of handwritten numeral recognition data set, the effect may not be very obvious, because it has been as high as 98%, but this is about manually adding modules. 
And the operation of class definition and incoming parameters is a positive exploration of the changes in the network structure. Questions and discussions are welcome.